added a button which creates a new cube
