+++ .gitignore
@@ -3,3 +3,4 @@
 
 # npm dependencies
 node_modules/
+package-lock.json
